refactor: move formatAddress function to its own file.
